requirements.txt
anthropic
openai
requests
python-dotenv
faiss-cpu
sentence-transformers
pandas
duckduckgo_search
gradio
## for langraph and langchain
langgraph>=0.0.25
pypdf>=3.15.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1